+++ src/main/kotlin/simplifiedclass/IDE.kt
@@ -2,10 +2,9 @@ package simplifiedclass
 
 import kotlin.collections.HashMap
 
-class IDE(hashMap: HashMap<String, Int>) {
+class IDE(hashMap: HashMap<String, Double>) {
 
     val countryList = hashMap.values.toList()
     val contPerson = hashMap.keys.toList()
-
 
 }
